+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/common/Checkable.kt
@@ -1,0 +1,8 @@
+package com.agileburo.anytype.core_ui.common
+
+/**
+ * Defines a view that can be checked
+ */
+interface Checkable {
+    val isChecked: Boolean
+}